Develop justin

+++ src/main/java/com/whydigit/efit/repo/AssetCategoryRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.whydigit.efit.entity.AssetCategoryVO;
 
 @Repository
-public interface AssetCategoryRepo extends JpaRepository<AssetCategoryVO, Integer> {
+public interface AssetCategoryRepo extends JpaRepository<AssetCategoryVO, Long> {
 	@Query(value = "select a from AssetCategoryVO a Where a.orgId=?1")
 	List<AssetCategoryVO> getAllAssetCategory(Long orgId);
